#206 [feat] : 활성화된 띠배너 조회 API를 분리한다
