+++ src/main/java/side/onetime/global/config/SecurityConfig.java
@@ -38,6 +38,7 @@ public class SecurityConfig {
             "/api/v1/urls/**",
             "/api/v1/tokens/**",
             "/api/v1/admin/**",
+            "/api/v1/banners/**",
             "/api/v1/users/onboarding",
             "/actuator/health"
     };
