+++ src/main/java/side/onetime/service/AdminService.java
@@ -302,12 +302,10 @@ public class AdminService {
      * 주어진 ID에 해당하며 현재 활성화 상태인 띠배너를 조회합니다.
      * - 해당 배너가 존재하지 않을 경우 예외가 발생합니다.
      *
-     * @param authorizationHeader 요청자의 액세스 토큰
      * @return 활성화된 배너 응답 객체
      */
     @Transactional(readOnly = true)
-    public GetActivatedBannerResponse getActivatedBanner(String authorizationHeader) {
-        jwtUtil.getAdminUserFromHeader(authorizationHeader);
+    public GetActivatedBannerResponse getActivatedBanner() {
         Banner banner = bannerRepository.findByIsActivatedTrue()
                 .orElseThrow(() -> new CustomException(AdminErrorStatus._NOT_FOUND_ACTIVATED_BANNER));
         return GetActivatedBannerResponse.from(banner);
