+++ src/test/java/side/onetime/admin/AdminControllerTest.java
@@ -581,53 +581,6 @@ public class AdminControllerTest extends ControllerTestConfig {
     }
 
     @Test
-    @DisplayName("활성화된 띠배너를 조회한다.")
-    public void getActivatedBanner() throws Exception {
-        // given
-        String accessToken = "Bearer test.jwt.token";
-        Long bannerId = 1L;
-        GetActivatedBannerResponse response = new GetActivatedBannerResponse(
-                bannerId, "공지사항", "Notice", "#FF5733", "#FFFFFF", true, "2025-04-01 12:00:00"
-        );
-
-        // when
-        Mockito.when(adminService.getActivatedBanner(any(String.class)))
-                .thenReturn(response);
-
-        // then
-        mockMvc.perform(RestDocumentationRequestBuilders.get("/api/v1/admin/banners/activated")
-                        .header("Authorization", accessToken))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.is_success").value(true))
-                .andExpect(jsonPath("$.code").value("200"))
-                .andExpect(jsonPath("$.message").value("활성화된 띠배너 조회에 성공했습니다."))
-                .andDo(MockMvcRestDocumentationWrapper.document("admin/banner-get-activated",
-                        preprocessRequest(prettyPrint()),
-                        preprocessResponse(prettyPrint()),
-                        resource(
-                                ResourceSnippetParameters.builder()
-                                        .tag("Admin API")
-                                        .description("활성화된 띠배너를 조회한다.")
-                                        .responseFields(
-                                                fieldWithPath("is_success").type(JsonFieldType.BOOLEAN).description("성공 여부"),
-                                                fieldWithPath("code").type(JsonFieldType.STRING).description("응답 코드"),
-                                                fieldWithPath("message").type(JsonFieldType.STRING).description("응답 메시지"),
-                                                fieldWithPath("payload").type(JsonFieldType.OBJECT).description("배너 정보"),
-                                                fieldWithPath("payload.id").type(JsonFieldType.NUMBER).description("배너 ID"),
-                                                fieldWithPath("payload.content_kor").type(JsonFieldType.STRING).description("한국어 내용"),
-                                                fieldWithPath("payload.content_eng").type(JsonFieldType.STRING).description("영어 내용"),
-                                                fieldWithPath("payload.background_color_code").type(JsonFieldType.STRING).description("배경 색상 코드"),
-                                                fieldWithPath("payload.text_color_code").type(JsonFieldType.STRING).description("텍스트 색상 코드"),
-                                                fieldWithPath("payload.is_activated").type(JsonFieldType.BOOLEAN).description("활성화 여부"),
-                                                fieldWithPath("payload.created_date").type(JsonFieldType.STRING).description("생성일자")
-                                        )
-                                        .responseSchema(Schema.schema("GetActivatedBannerResponse"))
-                                        .build()
-                        )
-                ));
-    }
-
-    @Test
     @DisplayName("띠배너를 전체 조회한다.")
     public void getAllBanners() throws Exception {
         // given
